Fix tile outlines disappearing
This fixes the issue when the page is full-sized and there's no 
overflow. When there is overflow, it fixes it only for the left tile, 
and not the right tile. But that's because of a long-standing CSS 
weirdness with Margins on the last items of things, so I'm ok with 
leaving it.

closes #49